Fix variables interpolation in validation request
