+++ src/specs/datasource.test.ts
@@ -57,48 +57,4 @@ describe('BigQueryDatasource', () => {
       expect(res).toEqual([]);
     });
   });
-
-  describe('When interpolating variables', () => {
-    beforeEach(() => {
-      ctx.variable = {};
-    });
-
-    describe('and value is a string', () => {
-      it('should return an unquoted value', () => {
-        expect(ctx.ds.interpolateVariable('abc', ctx.variable)).toEqual('abc');
-      });
-    });
-    describe('and value is a number', () => {
-      it('should return an unquoted value', () => {
-        expect(ctx.ds.interpolateVariable(1000, ctx.variable)).toEqual(1000);
-      });
-    });
-    describe('and value is an array of strings', () => {
-      it('should return comma separated quoted values', () => {
-        expect(ctx.ds.interpolateVariable(['a', 'b', 'c'], ctx.variable)).toEqual("'a','b','c'");
-      });
-    });
-
-    describe('and variable allows multi-value and is a string', () => {
-      it('should return a quoted value', () => {
-        ctx.variable.multi = true;
-        expect(ctx.ds.interpolateVariable('abc', ctx.variable)).toEqual("'abc'");
-      });
-    });
-
-    describe('and variable contains single quote', () => {
-      it('should return a quoted value', () => {
-        ctx.variable.multi = true;
-        expect(ctx.ds.interpolateVariable("a'bc", ctx.variable)).toEqual("'a''bc'");
-        expect(ctx.ds.interpolateVariable("a'b'c", ctx.variable)).toEqual("'a''b''c'");
-      });
-    });
-
-    describe('and variable allows all and is a string', () => {
-      it('should return a quoted value', () => {
-        ctx.variable.includeAll = true;
-        expect(ctx.ds.interpolateVariable('abc', ctx.variable)).toEqual("'abc'");
-      });
-    });
-  });
 });
